mfd: cros ec: spi: Add delay for asserting CS
Some ECs need a little time for waking up before they can accept SPI data at a
high speed. This is configurable via a DT property
"google,cros-ec-spi-pre-delay".

If this property isn't set, then no delay will be added. However, if set
it will cause a delay equal to the value passed to it to be inserted at
the beginning of a transaction.

Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>